myCode/classification/requirements_py3.7_tensor2.4_win64.txt
absl-py==0.15.0
aggdraw==1.3.16
astunparse==1.6.3
attrs==24.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
cachetools==5.5.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.4.0
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637091911212/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
entrypoints @ file:///C:/ci/entrypoints_1649926659561/work
et-xmlfile==1.1.0
fastjsonschema==2.21.1
flatbuffers==1.12
fonttools==4.38.0
gast==0.3.3
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.32.0
h5py==2.10.0
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel @ file:///C:/b/abs_21ykzkm7y_/croots/recipe/ipykernel_1662361803478/work
ipython @ file:///C:/Windows/TEMP/abs_45b5zb1l7q/croots/recipe/ipython_1659529855872/work
jedi @ file:///C:/ci/jedi_1644297241925/work
joblib==1.3.2
jsonschema==4.17.3
jupyter_client @ file:///C:/b/abs_d8fk_kz9zk/croot/jupyter_client_1676330195659/work
jupyter_core @ file:///C:/b/abs_b8r5lb5eus/croot/jupyter_core_1668084445779/work
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661934089728/work
nbformat==5.8.0
nest-asyncio @ file:///C:/b/abs_3a_4jsjlqu/croot/nest-asyncio_1672387322800/work
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.3
opt-einsum==3.3.0
packaging==24.0
pandas==1.1.5
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plotly==5.18.0
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
protobuf==3.20.3
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydot==2.0.0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing==3.1.4
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==305.1
pyzmq @ file:///C:/ci/pyzmq_1657597757589/work
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six==1.15.0
split-folders==0.5.1
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-gpu @ file:///F:/BaiduNetdiskDownload/TensorFlow%E5%AE%89%E8%A3%85%E6%95%99%E7%A8%8B/tensorflow_gpu-2.4.0-cp37-cp37m-win_amd64.whl#sha256=afc8998207d6442353737ce0aea7bb5248ad97618b1e24ad353725136d5a13e4
tensorflow-hub==0.16.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
threadpoolctl==3.1.0
tornado @ file:///C:/ci/tornado_1662476933490/work
tqdm==4.67.1
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
typing-extensions==3.7.4.3
urllib3==2.0.7
visualkeras==0.1.4
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
zipp==3.15.0
